+++ x/gov/types/errors.go
@@ -6,7 +6,6 @@ import (
 
 // x/gov module sentinel errors
 var (
-	ErrUnknownProposal       = errors.Register(ModuleName, 2, "unknown proposal")
 	ErrInactiveProposal      = errors.Register(ModuleName, 3, "inactive proposal")
 	ErrAlreadyActiveProposal = errors.Register(ModuleName, 4, "proposal already active")
 	// Errors 5 & 6 are legacy errors related to v1beta1.Proposal.
@@ -21,9 +20,7 @@ var (
 	ErrInvalidSigner           = errors.Register(ModuleName, 13, "expected gov account as only signer for proposal message")
 	ErrMetadataTooLong         = errors.Register(ModuleName, 15, "metadata too long")
 	ErrMinDepositTooSmall      = errors.Register(ModuleName, 16, "minimum deposit is too small")
-	ErrProposalNotFound        = errors.Register(ModuleName, 17, "proposal is not found")
 	ErrInvalidProposer         = errors.Register(ModuleName, 18, "invalid proposer")
 	ErrVotingPeriodEnded       = errors.Register(ModuleName, 20, "voting period already ended")
 	ErrInvalidProposal         = errors.Register(ModuleName, 21, "invalid proposal")
-	ErrVoteNotFound            = errors.Register(ModuleName, 23, "vote is not found")
 )